Partially rewrite MatrixHtmlParser to make it more correct.
We now always call open_element and close_element when opening/closing
tags, including when doing so from close_element (in case of malformed
HTML). Several bugs were fixed too (e.g. Element instances are truthy
when they have > 0 children so the check for falsiness has to be `is not
None`).